实体类同步更新

+++ yicode-question-openapi/yicode-question-openapi-api/src/main/java/com/yixihan/yicode/question/openapi/api/vo/request/question/AddQuestionAnswerReq.java
@@ -37,5 +37,5 @@ public class AddQuestionAnswerReq {
     private Integer answerTimeConsume;
     
     @ApiModelProperty(value = "内存消耗")
-    private Integer answerMemoryConsume;
+    private Double answerMemoryConsume;
 }
