+++ yicode-question-openapi/yicode-question-openapi-api/src/main/java/com/yixihan/yicode/question/openapi/api/vo/response/question/QuestionAnswerVO.java
@@ -49,7 +49,7 @@ public class QuestionAnswerVO {
     private Integer answerTimeConsume;
     
     @ApiModelProperty(value = "内存消耗")
-    private Integer answerMemoryConsume;
+    private Double answerMemoryConsume;
     
     @ApiModelProperty(value = "备注")
     private String answerNote;
